backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
google-generativeai==0.7.2
python-dotenv==1.0.1
sqlalchemy==2.0.30
pydantic==2.7.1
python-multipart==0.0.9
aiofiles==23.2.1
